add email, is_admin fields to users model

+++ db/migrate/20210305130616_create_notes.rb
@@ -4,7 +4,7 @@ class CreateNotes < ActiveRecord::Migration[6.1]
       t.references :user, null: false, foreign_key: true
       t.references :project, null: false, foreign_key: true
       t.references :category, null: false, foreign_key: true
-      t.integer :hours
+      t.float :hours
       t.string :date
 
       t.timestamps
